chore(utils): Renamed files using single syntax
